+++ backend/src/main/java/com/tecnocampus/LS2/protube_back/services/VideoService.java
@@ -1,52 +1,83 @@
 package com.tecnocampus.LS2.protube_back.services;
 
-import com.tecnocampus.LS2.protube_back.domain.Video;
-import com.tecnocampus.LS2.protube_back.persistence.VideoRepository;
+import com.tecnocampus.LS2.protube_back.domain.*;
+import com.tecnocampus.LS2.protube_back.persistence.*;
 import com.tecnocampus.LS2.protube_back.persistence.dto.VideoDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import java.net.URI;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class VideoService {
 
-    @Autowired
-    public VideoRepository videoRepository;
+    public final VideoRepository videoRepository;
+    private final UserRepository userRepository;
+    private final CategoryRepository categoryRepository;
+    private final TagRepository tagRepository;
+    private final CommentRepository commentRepository;
 
-    public VideoService() {
+    @Autowired
+    public VideoService(VideoRepository videoRepository, UserRepository userRepository, CategoryRepository categoryRepository, TagRepository tagRepository, CommentRepository commentRepository) {
+        this.videoRepository = videoRepository;
+        this.userRepository = userRepository;
+        this.categoryRepository = categoryRepository;
+        this.tagRepository = tagRepository;
+        this.commentRepository = commentRepository;
     }
 
     public List<VideoDTO> getVideos() {
-        return videoRepository.findAll().stream().map(this::toDTO).toList();
+        return videoRepository.findAll()
+                .stream()
+                .map(this::toDTO)
+                .toList();
     }
 
     public VideoDTO getVideoById(Long id) {
-        return videoRepository.findById(id).map(this::toDTO).orElse(null);
+        Video video = videoRepository.findById(id).orElse(null);
+        if (video == null) {
+            return null;
+        }
+        return toDTO(video);
     }
 
-    public URI createVideo(VideoDTO videoDTO) {
-        Video video = toEntity(videoDTO);
-        videoRepository.save(video);
-        return ServletUriComponentsBuilder.fromCurrentRequest()
-                .path("/{id}")
-                .buildAndExpand(video.getId())
-                .toUri();
+    public VideoDTO createVideo(VideoDTO videoDTO) {
+        Video video = toDomain(videoDTO);
+        Video saved = videoRepository.save(video);
+        return toDTO(saved);
     }
 
     public VideoDTO deleteVideo(Long id) {
         VideoDTO videoDTO = getVideoById(id);
-        videoRepository.deleteById(id);
+        if (videoDTO != null) {
+            videoRepository.deleteById(id);
+        }
         return videoDTO;
     }
 
     public VideoDTO updateVideo(Long id, VideoDTO videoDTO) {
-        Video video = toEntity(videoDTO);
-        video.setId(id);
-        videoRepository.save(video);
-        return toDTO(video);
+        Video video = videoRepository.findById(id).orElse(null);
+        if (video != null) {
+            video.updateVideo(videoDTO);
+            User user = userRepository.findById(videoDTO.userId()).orElse(null);
+            Category category = categoryRepository.findById(videoDTO.categoryId()).orElse(null);
+            Tag tag = tagRepository.findById(videoDTO.tagId()).orElse(null);
+            Comment comment = commentRepository.findById(videoDTO.commentId()).orElse(null);
+
+            if (user != null && category != null && tag != null && comment != null) {
+                video.getCategories().add(category);
+                video.getTags().add(tag);
+                video.getComments().add(comment);
+                video.setUser(user);
+
+            }
+
+            Video updated = videoRepository.save(video);
+            return toDTO(updated);
+        }
+        return null;
     }
 
     private VideoDTO toDTO(Video video) {
@@ -57,21 +88,22 @@ public class VideoService {
                 video.getHeight(),
                 video.getDuration(),
                 video.getDescription(),
+                video.getFilename(),
                 video.getUser() != null ? video.getUser().getId() : null,
-                video.getCategories() != null && !video.getCategories().isEmpty() ? video.getCategories().get(0).getId() : null,
-                video.getTags() != null ? video.getTags().stream().map(tag -> tag.getName()).toList() : null,
-                video.getComments() != null && !video.getComments().isEmpty() ? video.getComments().get(0).getId() : null
+                (video.getCategories() != null && !video.getCategories().isEmpty())
+                        ? video.getCategories().get(0).getId()
+                        : null,
+                (video.getTags() != null && !video.getTags().isEmpty())
+                        ? video.getTags().get(0).getId()
+                        : null,
+                (video.getComments() != null && !video.getComments().isEmpty())
+                        ? video.getComments().get(0).getId()
+                        : null
         );
     }
 
-    private Video toEntity(VideoDTO videoDTO) {
-        Video video = new Video();
-        video.setId(videoDTO.id());
-        video.setTitle(videoDTO.title());
-        video.setWidth(videoDTO.width());
-        video.setHeight(videoDTO.height());
-        video.setDuration(videoDTO.duration());
-        video.setDescription(videoDTO.description());
+    private Video toDomain(VideoDTO videoDTO) {
+        Video video = new Video(videoDTO);
         return video;
     }
 }
